requirements.txt
requests
pyyaml
markdown
pdfkit
PyPDF2
python-docx
openai